Prepare for initial DB migration

+++ MihuBot/MihuBot/DB/DbServiceCollectionExtensions.cs
@@ -5,21 +5,30 @@ namespace Microsoft.Extensions.DependencyInjection;
 
 public static class DbServiceCollectionExtensions
 {
+    private static string GetDatabasePath<TDBContext>() =>
+        typeof(TDBContext) == typeof(LogsDbContext) ? $"{Constants.StateDirectory}/MihuBot-logs.db" :
+        typeof(TDBContext) == typeof(MihuBotDbContext) ? $"{Constants.StateDirectory}/MihuBot.db" :
+        throw new NotSupportedException();
+
+
     public static void AddDatabases(this IServiceCollection services)
     {
-        services.AddPooledDbContextFactory<LogsDbContext>(options =>
-        {
-            options.UseSqlite($"Data Source={Constants.StateDirectory}/MihuBot-logs.db");
+        AddPooledDbContextFactory<LogsDbContext>(services);
+        AddPooledDbContextFactory<MihuBotDbContext>(services);
+    }
 
-            if (!OperatingSystem.IsLinux())
-            {
-                options.EnableSensitiveDataLogging();
-            }
-        });
+    public static async Task RunDatabaseMigrations(this IHost host)
+    {
+        await MigrateAsync<LogsDbContext>(host);
+        await MigrateAsync<MihuBotDbContext>(host);
+    }
 
-        services.AddPooledDbContextFactory<MihuBotDbContext>(options =>
+    private static void AddPooledDbContextFactory<TDbContext>(IServiceCollection services)
+        where TDbContext : DbContext
+    {
+        services.AddPooledDbContextFactory<TDbContext>(options =>
         {
-            options.UseSqlite($"Data Source={Constants.StateDirectory}/MihuBot.db");
+            options.UseSqlite($"Data Source={GetDatabasePath<TDbContext>()}");
 
             if (!OperatingSystem.IsLinux())
             {
@@ -28,28 +37,52 @@ public static class DbServiceCollectionExtensions
         });
     }
 
-    public static async Task RunDatabaseMigrations(this IHost host)
+    private static async Task MigrateAsync<TDbContext>(IHost host)
+        where TDbContext : DbContext
     {
-        using (var scope = host.Services.CreateScope())
+        using var scope = host.Services.CreateScope();
+
+        Console.WriteLine($"Applying {typeof(TDbContext).Name} migrations ...");
+
+        var factory = scope.ServiceProvider.GetRequiredService<IDbContextFactory<TDbContext>>();
+        await using var db = await factory.CreateDbContextAsync();
+
+        if (!(await db.Database.GetPendingMigrationsAsync()).Any())
         {
-            Console.WriteLine($"Applying {nameof(LogsDbContext)} migrations ...");
-
-            var factory = scope.ServiceProvider.GetRequiredService<IDbContextFactory<LogsDbContext>>();
-            await using var db = await factory.CreateDbContextAsync();
-
-            await db.Database.EnsureDeletedAsync();
-            await db.Database.MigrateAsync();
+            Console.WriteLine("No pending migrations");
+            return;
         }
 
-        using (var scope = host.Services.CreateScope())
+        string path = GetDatabasePath<TDbContext>();
+        string tempCopyPath = null;
+
+        if (OperatingSystem.IsWindows() && File.Exists(path))
         {
-            Console.WriteLine($"Applying {nameof(MihuBotDbContext)} migrations ...");
-
-            var factory = scope.ServiceProvider.GetRequiredService<IDbContextFactory<MihuBotDbContext>>();
-            await using var db = await factory.CreateDbContextAsync();
-
+            Console.WriteLine("DEBUG: Deleting existing database ...");
             await db.Database.EnsureDeletedAsync();
-            await db.Database.MigrateAsync();
+        }
+
+        if (File.Exists(path))
+        {
+            tempCopyPath = $"{Path.ChangeExtension(path, null)}-copy.tmp";
+
+            if (File.Exists(tempCopyPath))
+            {
+                throw new InvalidOperationException($"Backup copy already exists: {tempCopyPath}");
+            }
+
+            Console.WriteLine($"Creating a backup copy of {Path.GetFileName(path)}");
+            File.Copy(path, tempCopyPath, true);
+        }
+
+        await db.Database.MigrateAsync();
+
+        Console.WriteLine($"Migrated {typeof(TDbContext).Name}");
+
+        if (tempCopyPath is not null)
+        {
+            Console.WriteLine($"Deleting backup copy ({tempCopyPath})");
+            File.Delete(tempCopyPath);
         }
     }
 }
